code/backend/docker/app/requirements.txt
flask
flask_restful
numpy
onnxruntime

pandas 
bs4

elasticsearch
faiss-cpu
gensim

tqdm
requests
boto3
regex
sentencepiece
sacremoses

scikit-learn